refactor: Consolidate utility files for simpler architecture
High-impact restructuring to reduce complexity and improve maintainability:

CONSOLIDATED UTILITIES (24 files → 6 files):
- summary.ts: Merged summaryConstants, summaryGrouping, summaryHelpers
- filtering.ts: Merged filterUtils, viewFiltering, resultsUtils
- githubData.ts: Merged rawDataUtils, prEnrichment
- storage.ts: Merged storageUtils, indexedDB, usernameCache

BENEFITS:
- Fewer files to navigate (easier to find code)
- Logical grouping by domain (summary, filtering, data, storage)
- Cleaner imports (one import instead of many)
- Foundation for fast iterations

All imports updated across codebase and test files.
